refactor: GetBookmarkQueryParamsArgumentResolver 제거
