+++ src/main/java/com/meoguri/linkocean/configuration/WebConfiguration.java
@@ -11,8 +11,6 @@ import org.springframework.web.servlet.config.annotation.CorsRegistry;
 import org.springframework.web.servlet.config.annotation.ViewControllerRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
-import com.meoguri.linkocean.configuration.resolver.GetBookmarkQueryParamsArgumentResolver;
-
 import lombok.RequiredArgsConstructor;
 
 @Configuration
@@ -21,12 +19,9 @@ public class WebConfiguration implements WebMvcConfigurer {
 
 	private static final String ID = "id";
 	private static final int MAX_PAGE_SIZE = 8;
-	private final GetBookmarkQueryParamsArgumentResolver getBookmarkQueryParamsArgumentResolver;
 
 	@Override
 	public void addArgumentResolvers(final List<HandlerMethodArgumentResolver> resolvers) {
-		resolvers.add(getBookmarkQueryParamsArgumentResolver);
-
 		/* set default pageable as Page request [number:0, size 8, sort: id: DESC] */
 		SortHandlerMethodArgumentResolver sortArgumentResolver = new SortHandlerMethodArgumentResolver();
 		sortArgumentResolver.setFallbackSort(Sort.by(Sort.Direction.DESC, ID));
